+++ docs/userguide/src/en/bpmn/ch05-Spring.adoc
@@ -403,7 +403,7 @@ o.flowable.engine.impl.db.DbSqlSession   : performing create on engine with reso
 o.flowable.engine.impl.db.DbSqlSession   : performing create on history with resource org/flowable/db/create/flowable.h2.create.history.sql
 o.flowable.engine.impl.db.DbSqlSession   : performing create on identity with resource org/flowable/db/create/flowable.h2.create.identity.sql
 o.a.engine.impl.ProcessEngineImpl        : ProcessEngine default created
-o.a.e.i.a.DefaultAsyncJobExecutor        : Starting up the default async job executor [org.flowable.spring.SpringAsyncExecutor].
+o.a.e.i.a.DefaultAsyncJobExecutor        : Starting up the default async job executor [org.flowable.spring.job.service.SpringAsyncExecutor].
 o.a.e.i.a.AcquireTimerJobsRunnable       : {} starting to acquire async jobs due
 o.a.e.i.a.AcquireAsyncJobsDueRunnable    : {} starting to acquire async jobs due
 o.s.j.e.a.AnnotationMBeanExporter        : Registering beans for JMX exposure on startup
