+++ modules/flowable-job-spring-service/src/main/java/org/flowable/spring/job/service/SpringAsyncExecutor.java
@@ -10,8 +10,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.flowable.spring;
+package org.flowable.spring.job.service;
 
+import java.util.concurrent.RejectedExecutionException;
+
+import org.flowable.job.api.JobInfo;
+import org.flowable.job.service.impl.asyncexecutor.DefaultAsyncJobExecutor;
+import org.flowable.job.service.impl.asyncexecutor.ExecuteAsyncRunnable;
 import org.springframework.core.task.TaskExecutor;
 
 /**
@@ -24,31 +29,60 @@ import org.springframework.core.task.TaskExecutor;
  * </p>
  * 
  * @author Pablo Ganga
- * @deprecated use {@link org.flowable.spring.job.service.SpringAsyncExecutor}
  */
-@Deprecated
-public class SpringAsyncExecutor extends org.flowable.spring.job.service.SpringAsyncExecutor {
+public class SpringAsyncExecutor extends DefaultAsyncJobExecutor {
+
+    protected TaskExecutor taskExecutor;
+    protected SpringRejectedJobsHandler rejectedJobsHandler;
 
     public SpringAsyncExecutor() {
     }
 
     public SpringAsyncExecutor(TaskExecutor taskExecutor, SpringRejectedJobsHandler rejectedJobsHandler) {
-        super(taskExecutor, rejectedJobsHandler);
+        this.taskExecutor = taskExecutor;
+        this.rejectedJobsHandler = rejectedJobsHandler;
     }
 
-    @Override
+    public TaskExecutor getTaskExecutor() {
+        return taskExecutor;
+    }
+
+    /**
+     * Required spring injected {@link TaskExecutor} implementation that will be used to execute runnable jobs.
+     * 
+     * @param taskExecutor
+     */
+    public void setTaskExecutor(TaskExecutor taskExecutor) {
+        this.taskExecutor = taskExecutor;
+    }
+
     public SpringRejectedJobsHandler getRejectedJobsHandler() {
-        return (SpringRejectedJobsHandler) super.getRejectedJobsHandler();
+        return rejectedJobsHandler;
     }
 
     /**
      * Required spring injected {@link RejectedJobsHandler} implementation that will be used when jobs were rejected by the task executor.
      * 
      * @param rejectedJobsHandler
-     * @deprecated use {@link this#setRejectedJobsHandler(org.flowable.spring.job.service.SpringRejectedJobsHandler)}
      */
-    @Deprecated
     public void setRejectedJobsHandler(SpringRejectedJobsHandler rejectedJobsHandler) {
-        super.setRejectedJobsHandler(rejectedJobsHandler);
+        this.rejectedJobsHandler = rejectedJobsHandler;
     }
+
+    @Override
+    public boolean executeAsyncJob(JobInfo job) {
+        try {
+            taskExecutor.execute(new ExecuteAsyncRunnable(job, jobServiceConfiguration, jobEntityManager, asyncRunnableExecutionExceptionHandler));
+            return true;
+        } catch (RejectedExecutionException e) {
+            rejectedJobsHandler.jobRejected(this, job);
+            return false;
+        }
+    }
+
+    @Override
+    protected void initAsyncJobExecutionThreadPool() {
+        // Do nothing, using the Spring taskExecutor
+    }
+
 }
