+++ modules/flowable-spring/pom.xml
@@ -102,6 +102,10 @@
             <artifactId>flowable-spring-common</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.flowable</groupId>
+            <artifactId>flowable-job-spring-service</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-context</artifactId>
         </dependency>
