+++ modules/flowable-spring/src/test/resources/org/flowable/spring/test/components/SpringAsyncHistoryJobExecutorTest-context.xml
@@ -48,7 +48,7 @@
     <bean id="historyService" factory-bean="processEngine" factory-method="getHistoryService"/>
     <bean id="managementService" factory-bean="processEngine" factory-method="getManagementService"/>
 
-    <bean id="springAsyncExecutor" class="org.flowable.spring.SpringAsyncExecutor">
+    <bean id="springAsyncExecutor" class="org.flowable.spring.job.service.SpringAsyncExecutor">
         <property name="taskExecutor">
             <bean class="org.springframework.core.task.SimpleAsyncTaskExecutor" />
         </property>
@@ -56,7 +56,7 @@
         <property name="defaultAsyncJobAcquireWaitTimeInMillis" value="1000" />
     </bean>
     
-    <bean id="springAsyncHistoryExecutor" class="org.flowable.spring.SpringAsyncHistoryExecutor">
+    <bean id="springAsyncHistoryExecutor" class="org.flowable.spring.job.service.SpringAsyncHistoryExecutor">
         <property name="taskExecutor">
             <bean class="org.springframework.core.task.SimpleAsyncTaskExecutor" />
         </property>
