+++ pom.xml
@@ -504,6 +504,11 @@
                 <artifactId>flowable-job-service</artifactId>
                 <version>${project.version}</version>
             </dependency>
+            <dependency>
+                <groupId>org.flowable</groupId>
+                <artifactId>flowable-job-spring-service</artifactId>
+                <version>${project.version}</version>
+            </dependency>
 			<dependency>
 				<groupId>org.flowable</groupId>
 				<artifactId>flowable-camel</artifactId>
@@ -949,6 +954,7 @@
         <module>modules/flowable-task-service</module>
         <module>modules/flowable-job-service-api</module>
         <module>modules/flowable-job-service</module>
+        <module>modules/flowable-job-spring-service</module>
 		<module>modules/flowable-engine</module>
 	</modules>
